app/mqtt2db/requirements.txt
# requirements.txt

paho-mqtt           # Conn to MQTT
psycopg2-binary     # PostgreSQL
python-dotenv       # .env
